moves stepListStepData to its own file that can be imported in other sb stories
